requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
idna==3.10
pillow==11.2.1
razorpay==1.4.2
requests==2.32.4
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0

gunicorn